get weather by location (done)

+++ app/src/main/java/weather/budi/com/weatherapps/vo/TemperatureVO.java
@@ -13,25 +13,21 @@ public class TemperatureVO {
         this.temp = temp;
     }
 
-    public int getPressure() {
+    public double getPressure() {
         return pressure;
     }
 
-    public void setPressure(int pressure) {
+    public void setPressure(double pressure) {
         this.pressure = pressure;
     }
 
-    public int getHumidity() {
+    public double getHumidity() {
         return humidity;
     }
 
-    public void setHumidity(int humidity) {
+    public void setHumidity(double humidity) {
         this.humidity = humidity;
     }
-
-
-
-
 
     public double getTemp_min() {
         return temp_min;
@@ -51,8 +47,8 @@ public class TemperatureVO {
 
 
     private double temp;
-    private int pressure;
-    private int humidity;
+    private double pressure;
+    private double humidity;
     private double temp_min;
     private double temp_max;
 }
